+++ Prototype/src/World/Tile.cpp
@@ -6,10 +6,10 @@
 #include "World\objects\object.h"
 #include <iostream>
 
-const unsigned int Tile::TILE_SIZE = 16;
+const size_t Tile::TILE_SIZE = 16;
 //This is assuming a tilesheet is 512x512 pixels
-const unsigned int Tile::TILES_PER_ROW = 512 / Tile::TILE_SIZE;
-const unsigned int Tile::TILES_PER_COLUMN = 512 / Tile::TILE_SIZE;
+const size_t Tile::TILES_PER_ROW = 512 / Tile::TILE_SIZE;
+const size_t Tile::TILES_PER_COLUMN = 512 / Tile::TILE_SIZE;
 
 Tile::Tile(int _tileIndex, int _rotation, bool _flipH, bool _flipV, 
            std::optional<Polyshape> _polygon, int _objectIndex, std::vector<std::string> _objectparameters)
